Workaround for basic-ftp ftps connection issues
When a node basic-ftp client connects via ftps, the
readable function is getting called before the code
expects, and it immediately closes the socket.

Waiting for just a few ms before closing gets around
this issue.

This is not a root-cause fix, but it might be good
enough to use until time permits figuring out a
better fix.